Open an issue.

Guides through the process of creating an issue. Requires an user and an API key.

post_bug(text, title, component, ..., version, product, host)

Arguments

text
A character vector with the text of the bug you want to open.

title
A character vector with the title of the bug.

component
A character with the component of R you want to fill an issue with.

...
Named arguments passed to the API check documentation

version
A character of the Version you want to use eg "R 4.0.0".

product
A character of the product you want to use if missing "R" is automatically filled.

host
URL of the bugzilla instance if missing the R bugzilla is assumed.

Value

The ID of the issue posted.

See also